requirements.txt
chromadb==0.5.5
datasets==2.20.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.5
pandas==2.2.2
fastapi==0.112.2
uvicorn==0.30.6
